feat(output): separated the output file from the data aggregation
aggregators have one abstract parent class that shapes the behaviour,
each aggregator gets the current `build` plus a method to obtain the str
